+++ src/utils/initConsole.js
@@ -1,0 +1,6 @@
+// @flow
+
+import { addListener } from 'storyboard';
+import consoleListener from 'storyboard-listener-console';
+
+addListener(consoleListener);